+++ src/main/java/org/panteleyev/jpackage/DirectoryUtil.java
@@ -1,7 +1,5 @@
-/*
- Copyright © 2025 Petr Panteleyev <petr@panteleyev.org>
- SPDX-License-Identifier: BSD-2-Clause
- */
+// Copyright © 2025-2026 Petr Panteleyev
+// SPDX-License-Identifier: BSD-2-Clause
 package org.panteleyev.jpackage;
 
 import java.io.File;
@@ -10,22 +8,19 @@ import java.io.UncheckedIOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Comparator;
-import java.util.stream.Stream;
 
 final class DirectoryUtil {
 
     static boolean isNestedDirectory(Path parent, Path child) {
-        Path absoluteParent = parent.toAbsolutePath();
-        Path absoluteChild = child.toAbsolutePath();
+        var absoluteParent = parent.toAbsolutePath();
+        var absoluteChild = child.toAbsolutePath();
         return absoluteChild.startsWith(absoluteParent);
     }
 
     static void removeDirectory(Path dir) {
-        if (!dir.toFile().exists()) {
-            return;
-        }
+        if (!dir.toFile().exists()) return;
 
-        try (Stream<Path> paths = Files.walk(dir)) {
+        try (var paths = Files.walk(dir)) {
             paths.sorted(Comparator.reverseOrder())
                     .map(Path::toFile)
                     .forEach(File::delete);