+++ src/main/java/org/panteleyev/jpackage/Parameters.java
@@ -1,7 +1,5 @@
-/*
- Copyright © 2025 Petr Panteleyev <petr@panteleyev.org>
- SPDX-License-Identifier: BSD-2-Clause
- */
+// Copyright © 2025-2026 Petr Panteleyev
+// SPDX-License-Identifier: BSD-2-Clause
 package org.panteleyev.jpackage;
 
 import org.gradle.api.GradleException;
@@ -35,9 +33,7 @@ class Parameters {
     }
 
     public void addString(CommandLineParameter parameter, String value) {
-        if (value == null || value.isEmpty()) {
-            return;
-        }
+        if (value == null || value.isEmpty()) return;
 
         logger.info("  {} {}", parameter.getName(), value);
         params.add(parameter.getName());
@@ -45,10 +41,8 @@ class Parameters {
     }
 
     public void addString(CommandLineParameter parameter, Property<String> prop) {
-        String value = prop.getOrElse("");
-        if (value.isEmpty()) {
-            return;
-        }
+        var value = prop.getOrElse("");
+        if (value.isEmpty()) return;
 
         logger.info("  {} {}", parameter.getName(), value);
         params.add(parameter.getName());
@@ -56,25 +50,19 @@ class Parameters {
     }
 
     public void addBoolean(CommandLineParameter parameter, Property<Boolean> prop) {
-        if (!prop.getOrElse(false)) {
-            return;
-        }
+        if (!prop.getOrElse(false))            return;
 
         logger.info("  {}", parameter.getName());
         params.add(parameter.getName());
     }
 
     public void addFile(CommandLineParameter parameter, Property<String> prop, boolean mustExist) {
-        if (!prop.isPresent()) {
-            return;
-        }
+        if (!prop.isPresent()) return;
 
-        String value = prop.get();
-        if (value.isEmpty()) {
-            return;
-        }
+        var value = prop.get();
+        if (value.isEmpty()) return;
 
-        File file = projectDirectory.file(value).getAsFile();
+        var file = projectDirectory.file(value).getAsFile();
         if (mustExist && !file.exists()) {
             throw new GradleException("File or directory " + file.getAbsolutePath() + " does not exist");
         }
@@ -85,11 +73,9 @@ class Parameters {
     }
 
     public void addFile(CommandLineParameter parameter, RegularFileProperty prop, boolean mustExist) {
-        if (!prop.isPresent()) {
-            return;
-        }
+        if (!prop.isPresent()) return;
 
-        File file = prop.get().getAsFile();
+        var file = prop.get().getAsFile();
         if (mustExist && !file.exists()) {
             throw new GradleException("File or directory " + file.getAbsolutePath() + " does not exist");
         }
@@ -100,26 +86,9 @@ class Parameters {
     }
 
     public void addFile(CommandLineParameter parameter, DirectoryProperty prop, boolean mustExist) {
-        if (!prop.isPresent()) {
-            return;
-        }
+        if (!prop.isPresent()) return;
 
-        File file = prop.get().getAsFile();
-        if (mustExist && !file.exists()) {
-            throw new GradleException("File or directory " + file.getAbsolutePath() + " does not exist");
-        }
-
-        logger.info("  {} {}", parameter.getName(), file.getAbsolutePath());
-        params.add(parameter.getName());
-        params.add(file.getAbsolutePath());
-    }
-
-    public void addFile(CommandLineParameter parameter, String value, boolean mustExist) {
-        if (value == null || value.isEmpty()) {
-            return;
-        }
-
-        File file = projectDirectory.file(value).getAsFile();
+        var file = prop.get().getAsFile();
         if (mustExist && !file.exists()) {
             throw new GradleException("File or directory " + file.getAbsolutePath() + " does not exist");
         }