+++ src/test/java/org/panteleyev/jpackage/JPackageGradlePluginTest.java
@@ -1,10 +1,7 @@
-/*
- Copyright © 2022 Petr Panteleyev <petr@panteleyev.org>
- SPDX-License-Identifier: BSD-2-Clause
- */
+// Copyright © 2022-2026 Petr Panteleyev
+// SPDX-License-Identifier: BSD-2-Clause
 package org.panteleyev.jpackage;
 
-import org.gradle.api.Project;
 import org.gradle.testfixtures.ProjectBuilder;
 import org.junit.jupiter.api.Test;
 
@@ -13,9 +10,8 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 public class JPackageGradlePluginTest {
     @Test
     public void pluginRegistersTask() {
-        Project project = ProjectBuilder.builder().build();
+        var project = ProjectBuilder.builder().build();
         project.getPlugins().apply("org.panteleyev.jpackageplugin");
-
         assertNotNull(project.getTasks().findByName("jpackage"));
     }
 }